Read Bookshelf files.
Read an aux file of Bookshelf format and create Floorplan, Macros, Cells, Terminals, Nets and Pins.  Minor fix: Use doubles, instead of integers, to record area, because an area might be greater than the range of integer.  See Trello card 37 for details.

+++ src/cornersequence/CornerSequence.cpp
@@ -399,7 +399,7 @@ void CornerSequence::calculateEmptySpaceAreas() {
     cornerVerticalTilePlane->calculateEmptySpaceAreas();
 }
 
-int CornerSequence::getMacrosOccupiedRegionArea() {
+double CornerSequence::getMacrosOccupiedRegionArea() {
     return cornerVerticalTilePlane->getMacrosOccupiedRegionArea();
 }
